+++ scrapli_netconf/driver/async_driver.py
@@ -131,6 +131,25 @@ class AsyncNetconfScrape(AsyncScrape, NetconfScrapeBase):
         response._record_response(raw_response)  # pylint: disable=W0212
         return response
 
+    async def delete_config(self, target: str = "candidate") -> NetconfResponse:
+        """
+        Netconf delete-config operation
+
+        Args:
+            target: configuration source to target; startup|candidate
+
+        Returns:
+            NetconfResponse: scrapli_netconf NetconfResponse object
+
+        Raises:
+            N/A
+
+        """
+        response = self._pre_delete_config(target=target)
+        raw_response = await self.channel.send_input_netconf(response.channel_input)
+        response._record_response(raw_response)  # pylint: disable=W0212
+        return response
+
     async def commit(self) -> NetconfResponse:
         """
         Netconf commit config operation
@@ -224,6 +243,25 @@ class AsyncNetconfScrape(AsyncScrape, NetconfScrapeBase):
 
         """
         response = self._pre_rpc(filter_=filter_)
+        raw_response = await self.channel.send_input_netconf(response.channel_input)
+        response._record_response(raw_response)  # pylint: disable=W0212
+        return response
+
+    async def validate(self, source: str) -> NetconfResponse:
+        """
+        Netconf "validate" operation
+
+        Args:
+            source: configuration source to validate; typically one of running|startup|candidate
+
+        Returns:
+            NetconfResponse: scrapli_netconf NetconfResponse object
+
+        Raises:
+            N/A
+
+        """
+        response = self._pre_validate(source=source)
         raw_response = await self.channel.send_input_netconf(response.channel_input)
         response._record_response(raw_response)  # pylint: disable=W0212
         return response